requirements.txt
gym==0.15.4
pycolab==1.2
torch==1.9.0
matplotlib==3.1.2
pyyaml
